Clarify browser based Javascript vs Node
As I understand it, the "public nature" of javascript clients in the description here is due to the fact that the client secret would be exposed somewhere in a publicly accessible place. However, this is not the case for Node.js clients. Those clients will be running on the server side and are not necessarily required to be public. My proposed change attempts to clarify this a bit.